'plain' codec, start applying it in memory
It has become clear that we will need easily debuggable
memory models, even for the task of coding ssa memory
models.  This commit adds a new encoding interface, called
'plain', under an internal package and applies it to
some of the components of a memory model.

Also, some first tests ....