unified object refactoring (XL)
This change continues the unified object refactoring.

The new object builder is now in use in ssa2pal instead
of the old one.  This is still buggy, crashing on analyzing
its own source code, but atleast several packages are built
without crash, and atleast things compile.

results pkgres and related are not yet refactored.