ssa2pal: some work on tuples
tuples are weird, because they are not there at
runtime and so should not occupy memory like
structs do (which have their own distinct address).

Previously, I had tried to make tuples not be associated
with memory.Loc's nor objects.  This is too complicated.

This commit gives tuples their own address again.

It also adds some safety code in objects.Builder:
in order to ensure that the creation of a root compound
object triggers object creation (and not just typed
memory locations), we re-walk the object after creation
and reconstruct the object from the pal representation.

This walkObj method is still incomplete.