plain coding: add more robust test harness
We add a
 TestRoundTripClobber(c Coder, clob func(c Coder), verbose bool) error
function.

It encodes c, then calls clob(c) to clobber c and make it different
than it was before the encoding.  Then we decode c from the output
of the encoding.  then we re-encode c and return an error if the
strings resulting from the first and second encoding differ.