first draft: function calls
We have implemented function calls in a way which should facilitate
closures and interfaces a great deal.

rather than the traditional Anderson copy/transfer constraints
of arguments to parameters and function results to caller
local results, we

1. represent each function in contiguous memory as a sequence (p0, p1,
   ..., pn, r0, r1, ..., rm)  where each pi is a pointer to
an object of type of parameter i, and each ri is a pointer to an
object of type of return i.
2. upon analyzing a function we take *pi as the starting value of
the parameter.
3. Upon calling a function, we set *pi = argi and resi = *ri.

Now, given 2 functions f, g of the same signature, we have
copy/transfer constraints (fp0, fp1, ..., fpn, fr0, fr1, ..., frm) =
(gp0, gp1, ..., gpn, gr0, gr1, ..., grn).  then the points-to
of each parameter and result are transfered to f from g.

In this way, dynamic dispatch should be a cinch: just call the
function object associated with the node, it will have all
the loads/stores of any function assigned to it.

Something similar should work for interfaces:  we just keep a struct
of function objects defined as above, and on assignment to an interface
value, we assign the functions from the concrete type.

Likewise, this should facilitate closures by adding a free variables
component like the pi,ri components.